+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+
+### Changed
+
+
+### Fixed
+
+## [2.0.0-rc2] - 2020-06-25
 
 ### Added
 
@@ -146,7 +155,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use Java 11 inside container
 - Bump hibernate-validator from 6.0.13.Final to 6.1.0.Final
 
-[unreleased]: https://github.com/adorsys/keycloak-config-cli/compare/v2.0.0-rc1...HEAD
+[unreleased]: https://github.com/adorsys/keycloak-config-cli/compare/v2.0.0-rc2...HEAD
+[2.0.0-rc2]: https://github.com/adorsys/keycloak-config-cli/compare/v2.0.0-rc1...v2.0.0-rc2
 [2.0.0-rc1]: https://github.com/adorsys/keycloak-config-cli/compare/v1.4.1...v2.0.0-rc1
 [1.4.0]: https://github.com/adorsys/keycloak-config-cli/compare/v1.3.1...v1.4.0
 [1.3.1]: https://github.com/adorsys/keycloak-config-cli/compare/v1.3.0...v1.3.1